refactor(utils): give each misc.py and hydra.py helper its real owner
Second half of the ownership boundary #361 documented for utils/ ("only for
stable, domain-agnostic leaves; a function used by one owner stays with that
owner"). Both modules named here were listed there as staged migration
candidates.

misc.py mixed dynamic import, seeding, logging, memory and metrics. Four of its
six functions were dead repo-wide — set_seed, configure_logger, clear_memory and
flatten_dict have zero call sites in any file type; flatten_dict's only
reference was its own recursion. They are deleted rather than resplit. Dropping
set_seed does not lose reproducibility: the tree seeds explicitly per
torch.Generator (sde/noise.py, data/data_source.py, models/hunyuan_image3/seed.py,
train/backend/base_backend.py), which is the correct pattern across ranks and
actors; the global-state version it superseded was never called.

Of the two live functions, load_function had one consumer (config/validation.py,
via two function-local imports whose deferral bought nothing — misc.py imports
only numpy and torch, validation.py already imports torch at top level, and
there is no cycle), so it moves there and both imports become top-level.
aggregate_numeric_metrics has five call sites across train/, algorithms/ and
utils/, so it stays a utils leaf under a name that says what it is
(utils/metrics.py). misc.py is deleted.

hydra.py was the only thing making utils/ depend on unirl.distributed: two of
its lines wrap remote() from distributed/group/placement.py. All nine importers
are trainers (eight unirl/trainer/*, one experimental/refl), so it moves to
unirl/trainer/hydra.py. utils/ is a leaf again, which is what a future utils
rule in lint/check_core_dependencies.py would need.
