+++ unirl/trainer/eval_suites.py
@@ -9,7 +9,7 @@ from typing import Any, List, Optional
 from hydra.utils import instantiate
 from omegaconf import DictConfig, OmegaConf
 
-from unirl.utils.hydra import remote_hydra
+from unirl.trainer.hydra import remote_hydra
 
 logger = logging.getLogger(__name__)
