+++ unirl/utils/metrics.py
@@ -1,0 +1,33 @@
+"""Average numeric metrics across repeated metric dictionaries."""
+
+from typing import Any, Dict, List
+
+import torch
+
+
+def aggregate_numeric_metrics(metrics_list: List[Dict[str, Any]]) -> Dict[str, float]:
+    """Average numeric metric keys across repeated metric dictionaries."""
+    aggregated: Dict[str, float] = {}
+    if not metrics_list:
+        return aggregated
+
+    all_keys = set()
+    for metrics in metrics_list:
+        all_keys.update(metrics.keys())
+
+    for key in all_keys:
+        values: List[float] = []
+        for metrics in metrics_list:
+            if key not in metrics:
+                continue
+            value = metrics[key]
+            if isinstance(value, torch.Tensor):
+                value = value.item() if value.numel() == 1 else value.mean().item()
+            if isinstance(value, bool):
+                values.append(float(value))
+            elif isinstance(value, (int, float)):
+                values.append(float(value))
+        if values:
+            aggregated[key] = sum(values) / len(values)
+
+    return aggregated
